feat: APP-674 add stories for collaborators and members and remove from dashboard flow, name change dropdown, add missing space
