+++ web-marketplace/src/pages/Dashboard/MyProjects/MyProjects.tsx
@@ -3,27 +3,33 @@ import { useLocation, useNavigate } from 'react-router-dom';
 import { useLingui } from '@lingui/react';
 import { GeocodeFeature } from '@mapbox/mapbox-sdk/services/geocoding';
 import { Grid } from '@mui/material';
-import { useSetAtom } from 'jotai';
+import { useAtom, useSetAtom } from 'jotai';
 
 import { CreateProjectCard } from 'web-components/src/components/cards/CreateCards/CreateProjectCard';
 import ProjectCard from 'web-components/src/components/cards/ProjectCard';
-import { CogIcon } from 'web-components/src/components/icons/CogIcon';
 
+//import { CogIcon } from 'web-components/src/components/icons/CogIcon';
 import { useAuth } from 'lib/auth/auth';
 import { useTracker } from 'lib/tracker/useTracker';
 import { useWallet } from 'lib/wallet/wallet';
 
-import { projectsDraftState } from 'pages/ProjectCreate/ProjectCreate.store';
+import {
+  projectsCurrentStepAtom,
+  projectsDraftState,
+} from 'pages/ProjectCreate/ProjectCreate.store';
 import WithLoader from 'components/atoms/WithLoader';
 import { PostFlow } from 'components/organisms/PostFlow/PostFlow';
 
 import { useDashboardContext } from '../Dashboard.context';
 import { useFetchProjectByAdmin } from './hooks/useFetchProjectsByAdmin';
 import {
+  CREATE_POST,
+  CREATE_POST_DISABLED_TOOLTIP_TEXT,
   DRAFT_ID,
-  MANAGE_PROJECT_BUTTON_TEXT,
   MY_PROJECTS_BUTTON_TEXT,
   MY_PROJECTS_EMPTY_TITLE,
+  NOT_SUPPORTED_TOOLTIP_TEXT,
+  //MANAGE_PROJECT_BUTTON_TEXT,
 } from './MyProjects.constants';
 import {
   getDefaultProject,
@@ -39,6 +45,7 @@ const MyProjects = (): JSX.Element => {
   const { track } = useTracker();
   const { wallet, loginDisabled } = useWallet();
   const { activeAccountId, activeAccount } = useAuth();
+  const [projectsCurrentStep] = useAtom(projectsCurrentStepAtom);
 
   const { adminProjects, isLoadingAdminProjects } = useFetchProjectByAdmin({
     adminAccountId: activeAccountId,
@@ -92,7 +99,7 @@ const MyProjects = (): JSX.Element => {
                   isLoading={isLoadingAdminProjects}
                   variant="skeleton"
                 >
-                  <ProjectCard
+                  {/* <ProjectCard
                     asAdmin
                     adminPrompt={
                       sanityProfilePageData?.allProfilePage?.[0]
@@ -111,6 +118,51 @@ const MyProjects = (): JSX.Element => {
                     }}
                     track={track}
                     pathname={location.pathname}
+                  /> */}
+                  <ProjectCard
+                    asAdmin
+                    adminPrompt={
+                      sanityProfilePageData?.allProfilePage?.[0]
+                        ?.projectCardPromptRaw
+                    }
+                    {...getDefaultProject(!activeAccountId, _)}
+                    {...project}
+                    button={{
+                      text: _(CREATE_POST),
+                      disabled:
+                        !activeAccountId || project.draft || !project.location,
+                    }}
+                    onButtonClick={() => {
+                      setPostProjectId(project.id);
+                      setPostOffChainProjectId(project.offChainId);
+                      setPostProjectName(project.name);
+                      setPostProjectSlug(project.slug);
+                    }}
+                    onContainedButtonClick={() => {
+                      if (
+                        !project.offChain ||
+                        (project.offChain && project.published)
+                      ) {
+                        navigate(
+                          `/project-pages/${project.id}/edit/basic-info`,
+                        );
+                      } else {
+                        const currentStep = projectsCurrentStep[project?.id];
+                        navigate(
+                          `/project-pages/${project?.id}/${
+                            currentStep || 'basic-info'
+                          }`,
+                        );
+                      }
+                    }}
+                    track={track}
+                    pathname={location.pathname}
+                    createPostTooltipText={
+                      loginDisabled
+                        ? _(NOT_SUPPORTED_TOOLTIP_TEXT)
+                        : _(CREATE_POST_DISABLED_TOOLTIP_TEXT)
+                    }
+                    editProjectTooltipText={_(NOT_SUPPORTED_TOOLTIP_TEXT)}
                   />
                 </WithLoader>
               </Grid>
