+++ web-marketplace/src/pages/Dashboard/MyProjects/components/Collaborators.stories.tsx
@@ -1,0 +1,46 @@
+import React, { useState } from 'react';
+import type { Meta } from '@storybook/react';
+
+import { CollaboratorsManagement } from './Collaborators';
+import { mockCollaborators } from './Collaborators.mock';
+import { ProjectRoleType } from './Collaborators.types';
+
+const meta: Meta<typeof CollaboratorsManagement> = {
+  title: 'Dashboard/CollaboratorsManagement',
+  component: CollaboratorsManagement,
+  argTypes: {
+    collaborators: {
+      control: 'object',
+      description: 'List of collaborators',
+    },
+  },
+};
+export default meta;
+
+export const Default = (args: { collaborators: typeof mockCollaborators }) => {
+  const [collaborators, setCollaborators] = useState(args.collaborators);
+
+  const handleRoleChange = (id: string, role: ProjectRoleType) => {
+    setCollaborators(prev =>
+      prev.map(c => (c.id === id ? { ...c, projectRole: role } : c)),
+    );
+  };
+
+  const handleRemove = (id: string) => {
+    setCollaborators(prev => prev.filter(c => c.id !== id));
+  };
+
+  return (
+    <CollaboratorsManagement
+      {...args}
+      collaborators={collaborators}
+      onRoleChange={handleRoleChange}
+      onRemove={handleRemove}
+      onInvite={() => alert('Invite Collaborator')}
+    />
+  );
+};
+
+Default.args = {
+  collaborators: mockCollaborators,
+};
